+++ Candidato/correo.cpp
@@ -1,0 +1,7 @@
+#include "correo.h"
+
+void LeerCorreo(Correo &e, string u, string d){
+    e.user = u;
+    e.dom = d;
+
+}
