+++ Candidato/correo.h
@@ -1,0 +1,10 @@
+#pragma once
+#include<string>
+
+using namespace std;
+
+struct Correo {
+    string user;
+    string dom;
+};
+void LeerCorreo(Correo &, string, string);  //Leera el correo
