+++ VOTANTE/generadorAleatorio.h
@@ -1,0 +1,11 @@
+#pragma once
+#include <string>
+#include "constantes.h"
+
+using namespace std;
+
+
+int generarDNI();
+string generarDistrito();
+string generarNombre();
+string generarApellido();
